Refactor code

+++ ELearning/src/main/java/com/cybersoft/common/IndentifyUser.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import com.cybersoft.dto.UserDetailDto;
 //Class trả về id của user đang tương tác với hệ thống
 public class IndentifyUser {
-	public static int getIdPrincipal() {
+	public static Long getIdPrincipal() {
 		
 		UserDetailDto userDetail = (UserDetailDto) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 		return userDetail.getId();
