+++ ELearning/src/main/java/com/cybersoft/controller/UserTargetController.java
@@ -24,7 +24,7 @@ public class UserTargetController {
 	
 	//get Target thuộc về khóa học
 	@GetMapping("{id}")
-	public Object get(@PathVariable int id) {
+	public Object get(@PathVariable Long id) {
 		try {
 			List<TargetDto> dtos = targetService.getByCourse(id);
 			return new ResponseEntity<Object>(dtos, HttpStatus.ACCEPTED);
