+++ ELearning/src/main/java/com/cybersoft/repository/TargetRepository.java
@@ -18,5 +18,5 @@ public interface TargetRepository extends JpaRepository<Target, Integer> {
 	
 	//trả về list target thuộc về một khóa học
 	@Query("SELECT new com.cybersoft.dto.TargetDto(t.id, t.title,c.description) FROM Target t JOIN Course c ON t.courseId = c.id WHERE c.id = ?1")
-	public List<TargetDto> findTargetsByCourse(int id);
+	public List<TargetDto> findTargetsByCourse(Long id);
 }
